temp - this part 2 is killing me...
I'm trying to filter out solutions which cross the sides. Right now my
mistake is that the lines are made on the edge which is still valid
rather than just outside, so it seems like I invalidate everything
because of the corners or something.

Been working on this to long already so need a commit. Hopefully it
isn't to much work to move out the edge I check against.